backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.0
pydantic==2.7.0
